# docs/openapi/openapi.yaml
basePath: /
definitions:
  agentdaemon.remoteStreamPromptRequest:
    properties:
      generation:
        type: integer
      input:
        $ref: '#/definitions/connector.PromptInput'
    type: object
  agentdaemon.remoteSubmitPermissionRequest:
    properties:
      decision:
        $ref: '#/definitions/connector.PermissionDecision'
      generation:
        type: integer
    type: object
  agentdaemon.remoteSubmitPromptForUserChoiceRequest:
    properties:
      decision:
        $ref: '#/definitions/connector.PromptForUserChoiceDecision'
      generation:
        type: integer
    type: object
...
  connector.PermissionDecision:
    properties:
      approved:
        type: boolean
      by:
        description: user id
        type: string
      deliveryID:
        type: string
      deviceID:
        type: string
      note:
        type: string
      requestID:
        type: string
    type: object
  connector.PromptForUserChoiceDecision:
    properties:
      answers:
        items:
          type: string
        type: array
      by:
        type: string
      cancelled:
        type: boolean
      deliveryID:
        type: string
      deviceID:
        type: string
      questionAnswers:
        items:
          $ref: '#/definitions/connector.PromptForUserChoiceQuestionAnswer'
        type: array
      reason:
        type: string
      requestID:
        type: string
    type: object
  connector.PromptForUserChoiceQuestionAnswer:
    properties:
      answer:
        type: string
      answers:
        items:
          type: string
        type: array
      header:
        type: string
      questionID:
        type: string
    type: object
  connector.PromptInput:
    properties:
      agentConfig:
        additionalProperties: {}
        description: AgentConfig is the agent config (agents.config jsonb).
        type: object
      agentID:
        type: string
      agentName:
        type: string
      agentSlug:
        type: string
      conversationID:
        type: string
      conversationInitiatorID:
        type: string
      runID:
        description: |-
          RunID is the Parsar agent_run row id; the connector uses it
          to scope any per-run scratch state.
        type: string
      triggerAttachments:
        description: |-
          TriggerAttachments carries non-text payloads (today: images
          downloaded from a Feishu inbound message) that arrived alongside
          TriggerMessageContent. Connectors that don't understand multimodal
          input may safely ignore the field.
        items:
          $ref: '#/definitions/store.MessageAttachment'
        type: array
      triggerMessageContent:
        description: |-
          TriggerMessageContent is the user-facing message that drives this
          prompt. The connector decides how to fold it into its own prompt
          schema.
        type: string
      workspaceID:
        type: string
    type: object
...
  dev.createAgentBody:
    properties:
      capabilities:
        items:
          type: string
        type: array
      config:
        additionalProperties: {}
        type: object
      connector_type:
        type: string
      default_model_id:
        type: string
      description:
        type: string
      initial_capabilities:
        items:
          $ref: '#/definitions/dev.createAgentCapabilityBody'
        type: array
      inline_new_secrets:
        items:
          $ref: '#/definitions/dev.createAgentInlineSecretBody'
        type: array
      name:
        type: string
      runtime:
        type: string
      slug:
        type: string
      system_prompt:
        type: string
      visibility:
        type: string
    type: object
  dev.createAgentCapabilityBody:
    properties:
      capability_version_id:
        type: string
      configuration:
        additionalProperties: {}
        type: object
      pinning_mode:
        description: |-
          PinningMode is "latest" or "pinned". Empty falls back to store's
          default (pinned); create-agent dialog sends "latest" for new
          bindings unless the user picks a specific version.
        type: string
    type: object
  dev.createAgentInlineSecretBody:
    properties:
      display_name:
        type: string
      is_model:
        type: boolean
      kind:
        type: string
      plaintext:
        type: string
    type: object
...
  dev.updateAgentBody:
    properties:
      capabilities:
        items:
          type: string
        type: array
      config:
        additionalProperties: {}
        type: object
      connector_type:
        type: string
      default_model_id:
        type: string
      description:
        type: string
      inline_new_secrets:
        items:
          $ref: '#/definitions/dev.createAgentInlineSecretBody'
        type: array
      name:
        type: string
      slug:
        type: string
      system_prompt:
        type: string
      workspace_id:
        type: string
    type: object
...
  store.AgentSummary:
    properties:
      capabilities:
        items:
          type: string
        type: array
      config:
        additionalProperties: {}
        type: object
      connector_type:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
      visibility:
        type: string
      workspace_id:
        type: string
    type: object
...
  store.MessageAttachment:
    properties:
      data_base64:
        description: |-
          DataBase64 is the standard-base64 encoded raw bytes. Callers
          MUST NOT base64-encode again on persistence.
        type: string
      kind:
        type: string
      mime:
        description: |-
          MIME is the upstream-reported content type. Falls back to a
          kind-specific default when upstream omits it.
        type: string
      size:
        description: |-
          Size is the byte count before base64 expansion. Cached so
          observability surfaces don't have to decode.
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    Parsar harness API. Optional Feishu SSO and event delivery
    are enabled only when the corresponding PARSAR_FEISHU_* values are configured.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Parsar API
  version: 0.1.0
paths:
...
  /api/v1/agents/{agentID}:
    delete:
      description: Marks the agent as deleted; existing conversations are retained.
        Owner/admin only.
      operationId: deleteDevAgent
      parameters:
      - description: Agent UUID
        in: path
        name: agentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted agent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: agent_id must be a valid uuid
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Caller is not workspace owner/admin
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Unknown agent
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Soft-delete an agent
      tags:
      - agents
    patch:
      consumes:
      - application/json
      description: Applies a partial update. All fields are optional; nil pointers
        mean "leave as-is". Slug and runtime are immutable and rejected with 422.
      operationId: updateDevAgent
      parameters:
      - description: Agent UUID
        in: path
        name: agentID
        required: true
        type: string
      - description: Partial agent update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dev.updateAgentBody'
      produces:
      - application/json
      responses:
        "200":
          description: Updated agent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Malformed request body or invalid UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Caller is not workspace owner/admin
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Immutable field (slug/runtime), or unknown capability
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update mutable agent fields
      tags:
      - agents
...
  /api/v1/workspaces/{workspaceID}/agents:
    get:
      description: Returns agents the caller can address in the given workspace. By
        default only enabled/active agents; pass include_disabled=true for the admin
        view.
      operationId: listDevAgents
      parameters:
      - description: Workspace UUID
        in: path
        name: workspaceID
        required: true
        type: string
      - description: Include disabled/archived agents in the response
        in: query
        name: include_disabled
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Active workspace agents with profile basics
          schema:
            additionalProperties: true
            type: object
        "400":
          description: workspace_id must be a valid uuid
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Caller is not an active workspace member
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Database-backed read APIs are disabled
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List active agents in a workspace
      tags:
      - agents
    post:
      consumes:
      - application/json
      description: Creates an agent under the given workspace. Owner/admin only. inline_new_secrets
        are materialised into the shared secret vault before the agent is persisted;
        any binding entries in config are patched with the resolved ids.
      operationId: createDevAgent
      parameters:
      - description: Workspace UUID
        in: path
        name: workspaceID
        required: true
        type: string
      - description: Agent create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dev.createAgentBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created agent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: workspace_id must be a valid uuid, or body invalid
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Caller is not workspace owner/admin
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Immutable field or unknown capability
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an agent in a workspace
      tags:
      - agents
  /api/v1/workspaces/{workspaceID}/agents/{agentID}:
    get:
      description: Returns the complete agent summary and persisted config. The caller
        must be an active member of the workspace, and the agent must belong to that
        workspace.
      operationId: getDevAgent
      parameters:
      - description: Workspace UUID
        in: path
        name: workspaceID
        required: true
        type: string
      - description: Agent UUID
        in: path
        name: agentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Complete agent summary and config
          schema:
            $ref: '#/definitions/store.AgentSummary'
        "400":
          description: Invalid UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Caller is forbidden from accessing the workspace
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Workspace membership or agent not found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Database-backed read APIs are disabled
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an agent in a workspace
      tags:
      - agents
...
  /internal/agent-daemon/stream-prompt:
    post:
      consumes:
      - application/json
      description: Forwarded StreamPrompt entry point on the owner pod. Token-gated
        by the cluster-shared PARSAR_MASTER_KEY bearer.
      operationId: streamPromptInternal
      parameters:
      - description: Bearer <cluster shared token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: generation + PromptInput
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/agentdaemon.remoteStreamPromptRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: NDJSON stream of PromptEvent (application/x-ndjson)
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: stale owner
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Owner-pod stream-prompt (internal)
      tags:
      - internal
  /internal/agent-daemon/submit-permission:
    post:
      consumes:
      - application/json
      description: Forwarded SubmitPermission entry point on the owner pod. Token-gated
        by the cluster-shared bearer.
      operationId: submitPermissionInternal
      parameters:
      - description: Bearer <cluster shared token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: generation + PermissionDecision
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/agentdaemon.remoteSubmitPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"ok\":true}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: stale owner
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Owner-pod submit-permission (internal)
      tags:
      - internal
  /internal/agent-daemon/submit-prompt-for-user-choice:
    post:
      consumes:
      - application/json
      description: Forwarded SubmitPromptForUserChoice entry point on the owner pod.
        Token-gated by the cluster-shared bearer.
      operationId: submitPromptForUserChoiceInternal
      parameters:
      - description: Bearer <cluster shared token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: generation + PromptForUserChoiceDecision
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/agentdaemon.remoteSubmitPromptForUserChoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"ok\":true}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: stale owner
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Owner-pod submit-prompt-for-user-choice (internal)
      tags:
      - internal
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: cookie
    name: parsar_session
    type: apiKey
swagger: "2.0"